requirements.txt
fastapi>=0.68.0,<0.110.0
uvicorn>=0.15.0,<0.25.0
python-multipart>=0.0.5,<0.1.0
transformers>=4.20.0,<5.0.0
torch>=1.10.0,<3.0.0
torchvision>=0.11.0,<1.0.0
Pillow>=8.3.0,<11.0.0
pytest>=6.0.0,<8.0.0
requests>=2.25.0,<3.0.0
numpy>=1.21.0,<2.0.0
